# packages/api-contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "APMTRACK API",
    "version": "0.1.0",
    "description": "Contrat OpenAPI backend APMTRACK: sante, auth, RBAC, communes, utilisateurs, agents, zones, referentiel, PV, paiements, signalements, patrouilles, dashboard, audit logs et exports."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Developpement local"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/geography/import-csv": {
      "post": {
        "summary": "Importe le decoupage administratif national (SUPER_ADMIN)",
        "parameters": [
          {
            "name": "dry_run",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "text/csv": {
              "schema": {
                "type": "string",
                "description": "Une ligne par quartier. Colonnes: region, region_code, departement, departement_code, arrondissement, arrondissement_code, quartier, quartier_code, commune_code. Separateur ';' ou ',', BOM tolere. Creation et mise a jour uniquement: aucun DELETE."
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Rapport d'import",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeographyImportReport"
                }
              }
            }
          },
          "400": {
            "description": "En-tetes invalides ou fichier vide"
          },
          "403": {
            "description": "Reserve au SUPER_ADMIN"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
    "schemas": {
...
      "CsvRowError": {
        "type": "object",
        "properties": {
          "line": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "GeographyLevelCounts": {
        "type": "object",
        "properties": {
          "created": {
            "type": "integer"
          },
          "updated": {
            "type": "integer"
          },
          "matched": {
            "type": "integer"
          }
        }
      },
      "GeographyImportReport": {
        "type": "object",
        "properties": {
          "dry_run": {
            "type": "boolean"
          },
          "rows_total": {
            "type": "integer"
          },
          "regions": {
            "$ref": "#/components/schemas/GeographyLevelCounts"
          },
          "departements": {
            "$ref": "#/components/schemas/GeographyLevelCounts"
          },
          "arrondissements": {
            "$ref": "#/components/schemas/GeographyLevelCounts"
          },
          "quartiers": {
            "$ref": "#/components/schemas/GeographyLevelCounts"
          },
          "communes_linked": {
            "type": "integer"
          },
          "skipped": {
            "type": "integer"
          },
          "errors_total": {
            "type": "integer"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CsvRowError"
            }
          },
          "errors_truncated": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}